# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Yuso API","description":"Yuso API for transport","termsOfService":"None"},"paths":{"/api/Auction":{"get":{"tags":["Auction"],"summary":"Gets the auctions by user identifier.","operationId":"ApiAuctionGet","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Auction"],"summary":"Creates the or update auction.","operationId":"ApiAuctionPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"auctionModel","in":"body","description":"The auction model.","required":false,"schema":{"$ref":"#/definitions/AuctionModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Auction/{id}":{"get":{"tags":["Auction"],"summary":"Gets the auction by identifier.","operationId":"ApiAuctionByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"The identifier.","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/api/Auction/promoted":{"get":{"tags":["Auction"],"summary":"Gets the promoted auction.","operationId":"ApiAuctionPromotedGet","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}},"/filter":{"get":{"tags":["Auction"],"summary":"Gets the filtered auction.","operationId":"FilterGet","consumes":[],"produces":[],"parameters":[{"name":"search","in":"query","description":"The search.","required":false,"type":"string"},{"name":"toCity","in":"query","description":"To city.","required":false,"type":"string"},{"name":"fromCity","in":"query","description":"From city.","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/Auction/winner/{auctionId}":{"post":{"tags":["Auction"],"summary":"Sets the winner.","operationId":"ApiAuctionWinnerByAuctionIdPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"auctionId","in":"path","description":"The auction identifier.","required":true,"type":"string","format":"uuid"},{"name":"winner","in":"body","description":"The winner.","required":false,"schema":{"$ref":"#/definitions/WinnerModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Auction/win":{"get":{"tags":["Auction"],"operationId":"ApiAuctionWinGet","consumes":[],"produces":[],"parameters":[{"name":"status","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/Avatar/{id}":{"get":{"tags":["Avatar"],"summary":"Gets the avatar by identifier.","operationId":"ApiAvatarByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"The identifier.","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/api/Avatar":{"post":{"tags":["Avatar"],"summary":"Creates the or update avatar.","operationId":"ApiAvatarPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"avatarModel","in":"body","description":"The avatar model.","required":false,"schema":{"$ref":"#/definitions/AvatarModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Bid":{"post":{"tags":["Bid"],"summary":"Creates the bid.","operationId":"ApiBidPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"bidModel","in":"body","description":"The bid model.","required":false,"schema":{"$ref":"#/definitions/BidModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Bid/{auctionId}":{"get":{"tags":["Bid"],"summary":"Gets the bids with auction identifier.","operationId":"ApiBidByAuctionIdGet","consumes":[],"produces":[],"parameters":[{"name":"auctionId","in":"path","description":"The auction identifier.","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/login":{"post":{"tags":["JwtBearer"],"summary":"Logins the specified login model.","operationId":"LoginPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"loginModel","in":"body","description":"The login model.","required":false,"schema":{"$ref":"#/definitions/LoginModel"}}],"responses":{"200":{"description":"Success"}}}},"/register":{"post":{"tags":["JwtBearer"],"summary":"Registers the user.","operationId":"RegisterPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"registerModel","in":"body","description":"The register model.","required":false,"schema":{"$ref":"#/definitions/RegisterModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Profile":{"post":{"tags":["Profile"],"summary":"Creates the or update profile.","operationId":"ApiProfilePost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"registerModel","in":"body","description":"The register model.","required":false,"schema":{"$ref":"#/definitions/RegisterModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Profile/{id}":{"get":{"tags":["Profile"],"summary":"Gets the profile information by identifier.","operationId":"ApiProfileByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"The identifier.","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/api/User/password":{"post":{"tags":["User"],"summary":"Changes the password.","operationId":"ApiUserPasswordPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"passwordModel","in":"body","description":"The password model.","required":false,"schema":{"$ref":"#/definitions/ChangePasswordModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/User":{"get":{"tags":["User"],"summary":"Gets all users.","operationId":"ApiUserGet","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["User"],"summary":"Removes all users.","operationId":"ApiUserPost","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}}},"definitions":{"AuctionModel":{"description":"Auction Model.","type":"object","properties":{"name":{"description":"Gets or sets the name.","type":"string"},"fromCity":{"description":"Gets or sets from city.","type":"string"},"fromAddress":{"description":"Gets or sets from address.","type":"string"},"fromPostalCode":{"description":"Gets or sets from postal code.","type":"string"},"toCity":{"description":"Gets or sets to city.","type":"string"},"toAddress":{"description":"Gets or sets to address.","type":"string"},"toPostalCode":{"description":"Gets or sets to postal code.","type":"string"},"endDate":{"format":"date-time","description":"Gets or sets the end date.","type":"string"},"pickupOnDate":{"format":"date-time","description":"Gets or sets the pickup on date.","type":"string"},"deliverOnDate":{"format":"date-time","description":"Gets or sets the deliver on date.","type":"string"},"weight":{"format":"float","description":"Gets or sets the weight.","type":"number"},"aditionalDetails":{"description":"Gets or sets the details.","type":"string"},"pickupGeo":{"description":"Gets or sets the pickup geo.","type":"string"},"deliveryGeo":{"description":"Gets or sets the delivery geo.","type":"string"},"distance":{"format":"float","description":"Gets or sets the distance.","type":"number"}}},"WinnerModel":{"description":"Winner Model.","type":"object","properties":{"winnerId":{"format":"uuid","description":"Gets or sets the winner identifier.","type":"string"}}},"AvatarModel":{"description":"Avatar Model.","type":"object","properties":{"image":{"format":"byte","description":"Gets or sets the image.","type":"string"}}},"BidModel":{"description":"Bid model.","type":"object","properties":{"price":{"format":"int32","description":"Gets or sets the price.","type":"integer"},"auctionId":{"format":"uuid","description":"Gets or sets the auction identifier.","type":"string"}}},"LoginModel":{"description":"Login model.","type":"object","properties":{"email":{"description":"Gets or sets the email.","type":"string"},"password":{"description":"Gets or sets the password.","type":"string"}}},"RegisterModel":{"description":"Register model.","type":"object","properties":{"firstName":{"description":"Gets or sets the first name.","type":"string"},"lastName":{"description":"Gets or sets the last name.","type":"string"},"address":{"description":"Gets or sets the address.","type":"string"},"city":{"description":"Gets or sets the city.","type":"string"},"county":{"description":"Gets or sets the county.","type":"string"},"postalCode":{"description":"Gets or sets the postal code.","type":"string"},"phone":{"description":"Gets or sets the phone.","type":"string"},"password":{"description":"Gets or sets the password.","type":"string"},"email":{"description":"Gets or sets the email.","type":"string"}}},"ChangePasswordModel":{"description":"Password Model.","type":"object","properties":{"oldPassword":{"description":"Gets or sets the old password.","type":"string"},"newPassword":{"description":"Gets or sets the new password.","type":"string"}}}},"securityDefinitions":{"Bearer":{"name":"Authorization","in":"header","type":"apiKey","description":"JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""}},"security":[{"Bearer":[]}]}